rag/backup/server/requirements.txt
fastmcp[cli]
psycopg2-binary
pgvector
sentence-transformers
python-dotenv
pydantic
uvicorn